Completely implement the GetRoom function
Implement the map class function that returns the position of a random room in the dungeon, so enemy placement can occur. It first traverses the bsp tree and, if the room is the starting room, returns the first one. Else, it returns a random item in the rest of the nodes