Implement a staircase and level-change mechanic
Each map holds the location of the staircase to the next level in it and draws it when render() is called. The Engine class now has a self-contained NewLevel() function that handles the creation of a new level and the deletion of all the entities in the old level. NewLevel is called whenever the player ends a turn occupying the tile where the staircase is.